Move getMorher() to the Unicorn namespace
Make it possible to use as Unicorn.getMorpher() in the minimized script.
